# contracts/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Odoo Operations",
    "version": "0.2.0"
  },
  "paths": {
    "/v1/odoo/info": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Info",
        "operationId": "info_v1_odoo_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/customers": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Customers",
        "operationId": "customers_v1_customers_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "revision",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/catalog": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Catalog",
        "operationId": "catalog_v1_catalog_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/opportunities": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Opportunities",
        "operationId": "opportunities_v1_opportunities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/opportunities/{opportunity_id}": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Opportunity",
        "operationId": "opportunity_v1_opportunities__opportunity_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "opportunity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Opportunity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/quotes/prepare": {
      "post": {
        "tags": [
          "business-v1"
        ],
        "summary": "Quote",
        "operationId": "quote_v1_quotes_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/activities/prepare": {
      "post": {
        "tags": [
          "business-v1"
        ],
        "summary": "Activity",
        "operationId": "activity_v1_activities_prepare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/proposals/{proposal_id}": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Get Proposal",
        "operationId": "get_proposal_v1_proposals__proposal_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals/{proposal_id}/approve": {
      "post": {
        "tags": [
          "business-v1"
        ],
        "summary": "Approve",
        "operationId": "approve_v1_proposals__proposal_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proposals/{proposal_id}/execute": {
      "post": {
        "tags": [
          "business-v1"
        ],
        "summary": "Execute",
        "operationId": "execute_v1_proposals__proposal_id__execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "proposal_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Proposal Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/{operation_id}": {
      "get": {
        "tags": [
          "business-v1"
        ],
        "summary": "Status",
        "operationId": "status_v1_operations__operation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/operations/{operation_id}/retry": {
      "post": {
        "tags": [
          "business-v1"
        ],
        "summary": "Retry",
        "operationId": "retry_v1_operations__operation_id__retry_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "operation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Operation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActivityInput": {
        "properties": {
          "opportunity_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Opportunity Id"
          },
          "assignee_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Assignee Id"
          },
          "due_date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "title": "Due Date"
          },
          "summary": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "pattern": "^[^<>]+$",
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "opportunity_id",
          "assignee_id",
          "due_date",
          "summary"
        ],
        "title": "ActivityInput"
      },
...
      "ApprovalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "proposal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Proposal Id"
          },
          "approver_id": {
            "type": "string",
            "format": "uuid",
            "title": "Approver Id"
          },
          "payload_hash": {
            "type": "string",
            "title": "Payload Hash"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "proposal_id",
          "approver_id",
          "payload_hash",
          "expires_at",
          "created_at"
        ],
        "title": "ApprovalOut"
      },
      "ApproveInput": {
        "properties": {
          "payload_hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "title": "Payload Hash"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "payload_hash"
        ],
        "title": "ApproveInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProposalOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Actor Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "quote",
              "activity"
            ],
            "title": "Kind"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "preview": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/QuotePreview"
              },
              {
                "$ref": "#/components/schemas/ActivityPreview"
              }
            ],
            "title": "Preview",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "activity": "#/components/schemas/ActivityPreview",
                "quote": "#/components/schemas/QuotePreview"
              }
            }
          },
          "payload_hash": {
            "type": "string",
            "title": "Payload Hash"
          },
          "contract_version": {
            "type": "string",
            "const": "1.0",
            "title": "Contract Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "actor_id",
          "kind",
          "payload",
          "preview",
          "payload_hash",
          "contract_version",
          "created_at",
          "expires_at"
        ],
        "title": "ProposalOut"
      },
      "QuoteInput": {
        "properties": {
          "customer_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Customer Id"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Item"
            },
            "type": "array",
            "maxItems": 20,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "customer_id",
          "items"
        ],
        "title": "QuoteInput"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}